Use `process.execPath` instead of `node` executable.
When this task is run with a different node.js version than the "node"
executable in the path (or even worse, if no "node" is in the path), we have
a version mismatch, which leads to strange effects if we are dealing with
compiled modules.

So it is better to use `process.execPath` which should point to the
currently running version of node.